Handling Beacons and Proximity Event moved from ContextManager to new class ProximityHandler + code refactoring
