+++ AlpsSDK/AlpsSDK.swift
@@ -88,8 +88,6 @@ protocol AlpsSDK {
     func stopUpdatingLocation()
     func startRangingBeacons(forUuid: UUID, identifier: String)
     func stopRangingBeacons(forUuid: UUID)
-    func startBeaconsProximityEvent(forCLProximity: CLProximity)
-    func stopBeaconsProximityEvent(forCLProximity: CLProximity)
     
     // Beacons
     func getClosestOnBeaconUpdate(completion: @escaping ((_ beacon: CLBeacon) -> Void))
